+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/grouping/GroupingAction.java
@@ -2,7 +2,6 @@ package org.jhotdraw.draw.action.grouping;
 
 import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.draw.action.AbstractSelectedAction;
-import org.jhotdraw.draw.action.grouping.strategy.GroupingStrategy;
 import org.jhotdraw.draw.figure.CompositeFigure;
 import org.jhotdraw.util.ResourceBundleUtil;
