+++ jhotdraw-gui/src/main/java/org/jhotdraw/gui/action/ButtonFactory.java
@@ -22,8 +22,8 @@ import org.jhotdraw.draw.action.arrange.ArrangeAction;
 import org.jhotdraw.draw.action.arrange.ArrangeBringToFrontStrategy;
 import org.jhotdraw.draw.action.arrange.ArrangeSendToBackStrategy;
 import org.jhotdraw.draw.action.grouping.GroupingActionFactory;
-import org.jhotdraw.draw.action.grouping.strategy.GroupStrategy;
-import org.jhotdraw.draw.action.grouping.strategy.UngroupStrategy;
+import org.jhotdraw.draw.action.grouping.GroupStrategy;
+import org.jhotdraw.draw.action.grouping.UngroupStrategy;
 import org.jhotdraw.draw.decoration.ArrowTip;
 import org.jhotdraw.draw.decoration.LineDecoration;
 import org.jhotdraw.draw.event.SelectionComponentRepainter;
