+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/SVGApplicationModel.java
@@ -32,8 +32,8 @@ import org.jhotdraw.draw.action.arrange.ArrangeAction;
 import org.jhotdraw.draw.action.arrange.ArrangeBringToFrontStrategy;
 import org.jhotdraw.draw.action.arrange.ArrangeSendToBackStrategy;
 import org.jhotdraw.draw.action.grouping.GroupingActionFactory;
-import org.jhotdraw.draw.action.grouping.strategy.GroupStrategy;
-import org.jhotdraw.draw.action.grouping.strategy.UngroupStrategy;
+import org.jhotdraw.draw.action.grouping.GroupStrategy;
+import org.jhotdraw.draw.action.grouping.UngroupStrategy;
 import org.jhotdraw.draw.io.InputFormat;
 import org.jhotdraw.draw.io.OutputFormat;
 import org.jhotdraw.gui.JFileURIChooser;
