+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/gui/ActionsToolBar.java
@@ -25,8 +25,8 @@ import org.jhotdraw.action.edit.SelectAllAction;
 import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.draw.action.*;
 import org.jhotdraw.draw.action.grouping.GroupingActionFactory;
-import org.jhotdraw.draw.action.grouping.strategy.GroupStrategy;
-import org.jhotdraw.draw.action.grouping.strategy.UngroupStrategy;
+import org.jhotdraw.draw.action.grouping.GroupStrategy;
+import org.jhotdraw.draw.action.grouping.UngroupStrategy;
 import org.jhotdraw.gui.JPopupButton;
 import org.jhotdraw.gui.action.ButtonFactory;
 import org.jhotdraw.gui.plaf.palette.PaletteButtonUI;
